requirements.txt
altgraph==0.17.4
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PySocks==1.7.1
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
selenium==4.16.0
setuptools==69.0.3
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
